prod/requirements.txt
torch
streamlit
torchvision
numpy
pandas
pillow
